+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added warnings in FrameSettings inspector when using DXR and atempting to use Asynchronous Execution.
 - Exposed Stencil bits that can be used by the user.
 - Added history rejection based on velocity of intersected objects for directional, point and spot lights.
+- Added a affectsVolumetric field to the HDAdditionalLightData API to know if light affects volumetric fog.
 
 ### Fixed
 - Sorting, undo, labels, layout in the Lighting Explorer.